Added database header read/write support

+++ kotpass/src/main/kotlin/io/github/anvell/kotpass/models/FormatVersion.kt
@@ -1,10 +1,10 @@
 package io.github.anvell.kotpass.models
 
 data class FormatVersion(
-    val major: Int,
-    val minor: Int
+    val major: Short,
+    val minor: Short
 ) {
-    fun isAtLeast(major: Int, minor: Int): Boolean {
+    fun isAtLeast(major: Short, minor: Short): Boolean {
         return this.major > major || (this.major == major && this.minor >= minor)
     }
 }
